requirements.txt
Django==1.9
djangorestframework==3.3.1
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.7.2
MySQL-python==1.2.5
PyJWT==1.4.0
wheel==0.24.0
